o tested aggreaget_ra.cc & _hc.cc o fixed a bug in array_copy.cc o added tracing calls to finish & reduce methods
git-svn-id: svn://svn.code.sf.net/p/x10/code@4612 a326200e-df2c-42e0-81fa-597e909af41d

+++ trunk/x10.lib/src/aggregate_ra.cc
@@ -1,7 +1,7 @@
 /*
  * (c) Copyright IBM Corporation 2007
  *
- * $Id: aggregate_ra.cc,v 1.6 2007-10-24 07:16:31 ganeshvb Exp $
+ * $Id: aggregate_ra.cc,v 1.7 2007-10-24 09:51:48 ganeshvb Exp $
  * This file is part of X10 Runtime System.
  */
 
@@ -216,7 +216,7 @@ namespace x10lib {
 
   //template <typename FUNC>  
   x10_err_t 
-  asyncFlush_ra (x10_async_handler_t hndlr, size_t size, char* data, int len, int log_table_size,
+  asyncFlush_ra (x10_async_handler_t hndlr, size_t size, char* data, int len, long log_table_size,
                  int place_id_mask)
   {
     X10_DEBUG (1,  "Entry");
